requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
mcap==1.1.1
mcap-ros2-support==0.5.2
numpy==2.1.3
anthropic==0.40.0
pydantic==2.10.2